+++ index.js
@@ -2,8 +2,7 @@ import closest from 'dom-closest';
 
 export default ( selector, listener ) => ( e ) => {
 	const isDelegated = (
-		closest(e.target, selector) === e.target &&
-		e.currentTarget.contains(e.target) &&
+		e.currentTarget.contains(closest(e.target, selector)) &&
 		e.currentTarget !== e.target
 	);
 	if ( isDelegated ) {